Revert "Revert "Added short delay before starting script""
This reverts commit 5395cec0d620540cc6be983b85c9973c9cf45853.

+++ src/main/java/org/lolwat/Core.java
@@ -100,7 +100,7 @@ public class Core implements Runnable {
         jsonOutput.append("\"BB_DISPLAYNAME\": \"").append(displayName).append("\", ");
         jsonOutput.append("\"BB_TYPE\": \"").append(accountType).append("\", ");
         jsonOutput.append("\"BB_WORLD\": ").append(world).append(", ");
-        jsonOutput.append("\"BB_GP\": ").append(lastTotalGP).append(", ");
+        jsonOutput.append("\"BB_GP\": ").append((long)lastTotalGP).append(", ");
         jsonOutput.append("\"BB_TTL\": ").append(lastTotalLevel).append(", ");
         jsonOutput.append("\"BB_QP\": ").append(lastQuestPoints).append(", ");
         jsonOutput.append(getStatsJson());
